refactor(controller): main 함수의 Contorller 로직을 별도 클래스로 분리
MVC 패턴 적용을 위해 Application.kt의 main 함수에 포함되어있는 전체 실행
흐름 관리 로직(입력 → 검증 → 경주 → 출력)을 RacingController 클래스로 분리했습니다.

이를 통해 main 함수는 프로그램 실행의 진입 역할만 담당하고,
비즈니스 로직은 Controller가 관리하도록 구조를 개선했습니다.

View는 단순히 입출력만 담당하므로 싱글톤 객체로 충분하다고 생각하여
object를 사용해 구현하였지만, Controller는 각 실행마다 상태를 관리하고,
경주 진행 과정에서 데이터를 조합해야 하므로 인스턴스 생성이 필요한 class로 구현했습니다.
